model definition and editing for artist group

+++ server/routes/delete.js
@@ -2,6 +2,7 @@
 	'use strict';
 
 	var Promise = require('bluebird');
+	var Sequelize = require('sequelize');
 
 	module.exports = function (router, models) {
 		router.delete('/api/artist/:_id', function (req, res) {
@@ -9,6 +10,7 @@
 			var promises = [];
 			var songArtistCount = 0;
 			var albumArtistCount = 0;
+			var artistGroupCount = 0;
 
 			promises.push(models.SongArtist.count({
 					where: { ArtistId: id } 
@@ -22,9 +24,17 @@
 				albumArtistCount = count;
 			}));
 
+			promises.push(models.ArtistGroup.count({
+					where: Sequelize.or({ GroupId: id }, { MemberId: id })
+			}).then(function (count) {
+				artistGroupCount = count;
+			}));
+
 			Promise.all(promises)
 			.then(function () {
-				if (songArtistCount === 0 && albumArtistCount === 0) {
+				if (songArtistCount === 0 &&
+						albumArtistCount === 0 &&
+					  artistGroupCount === 0) {
 					models.Artist.destroy({
 						where: { id: id }
 					})
